requirements.txt
telethon
python-dotenv
nest_asyncio
flask
